+++ battlecode22-scaffold/autoscript.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 
 team1=ASageActionBot
-team2=OTempBot
-# team2=OCommMiningBot
+team2=OCommMiningBot
 
 mkdir -p logs
 if test -f "logs/results.log"; then